+++ app/domain/authentication/authn_jwt/validate_and_decode/validate_and_decode_token.rb
@@ -3,7 +3,7 @@ module Authentication
     module ValidateAndDecode
       # ValidateAndDecodeToken command class is responsible to validate the JWT token 2 times:
       # 1st we are validating only the signature.
-      # 2nd we are validating the claims, by checking the token content to decide which claims are mandatory
+      # 2nd we are validating the claims, by checking the token content to decide which claims are enforced
       # for the 2nd validation
       ValidateAndDecodeToken ||= CommandClass.new(
         dependencies: {
@@ -96,7 +96,7 @@ module Authentication
           claims_to_validate.each do |jwt_claim|
             claim_name = jwt_claim.name
             if @decoded_token[claim_name].blank?
-              raise Errors::Authentication::AuthnJwt::MissingMandatoryClaim, claim_name
+              raise Errors::Authentication::AuthnJwt::MissingEnforcedClaim, claim_name
             end
 
             verification_option = @get_verification_option_by_jwt_claim.call(jwt_claim: jwt_claim)